requirements.txt
torch
torchvision
tqdm
opencv-python
numpy
pandas
pyyaml
